ReadInterpolate: initial functional commit
ReadInterpolate is a FileReader like thorn that uses LocalInterpUniform to
interpolate the read in data onto the target grid. This means that the
grid structures can be completely arbirary. However only interpolators
provided by LocalInterp or AEILocalInterp can be used. This means
essentially only Lagrange and Hermite polynomials but no ENO.

Right now the reader will only allow for the selection of which grid
variables to read via a global regular expresion a la hdf5_slicer. It is
limited to CCTK_REAL grid functions as well.

For a usage example see the test admhydro.